STORE-784 Updating User UI Working on Search results
