+++ server/openstorefront/openstorefront-core/model/src/main/java/edu/usu/sdl/openstorefront/core/view/MultipleIds.java
@@ -24,13 +24,13 @@ import java.util.List;
  * Use to consume multiple ids
  * @author dshurtleff
  */
-public class MulitpleIds
+public class MultipleIds
 {
 	@ConsumeField
 	@DataType(String.class)		
 	private List<String> ids = new ArrayList<>();
 
-	public MulitpleIds()
+	public MultipleIds()
 	{
 	}
