+++ server/openstorefront/openstorefront-web/src/main/webapp/client/css/app.css
@@ -25,6 +25,10 @@ and open the template in the editor.
 		overflow: visible;
 }
 
+.action-icon {
+	font-size: 16px;
+}
+
 .link {
     color: #3C3B3B;;
     text-decoration: none;	
@@ -136,6 +140,8 @@ and open the template in the editor.
 	background-color: white;
 	background-image: none;
 	border: 1px solid #cccccc;
+	border-left: 0px;
+    border-right: 0px;	
 	-webkit-box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
 	box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075);
 	-webkit-transition: border-color ease-in-out 0.15s, box-shadow ease-in-out 0.15s;
@@ -165,12 +171,14 @@ and open the template in the editor.
 	background: white;
     padding: 10px;
 	border: 1px solid lightgray;
+	margin-top: -1px;
 }
 
 .searchresults-item:nth-child(even){
 	background: whitesmoke;
     padding: 10px;
 	border: 1px solid lightgray;
+	margin-top: -1px;
 }
 
 .searchresults-item-update{
@@ -178,9 +186,23 @@ and open the template in the editor.
 	color: darkgray;
 }
 
+.searchresults-item-org {
+	font-size: 12px;
+	color: grey;
+}
+
 .searchresults-item:hover{
 	background: #d8e9f6;
 	cursor: pointer;
+}
+
+.searchresults-tag{
+	border: 1px solid lightgray;
+	border-radius: 5px 5px 5px 5px;
+	background: rgba(68,30,90,1);
+	color: white;
+	font-weight: bold;
+	padding: 5px;
 }
 
 .details-title-name {
@@ -427,8 +449,7 @@ tr.details-table:hover {
 }
 
 .home-highlight-item-desc {
-	padding-left: 5px;
-	padding-right: 5px;
+	padding: 5px;	
 }
 
 .home-highlight-approved {
